+++ src/nano.h
@@ -578,7 +578,7 @@ enum
 #endif /* !NANO_TINY */
 
 /* The maximum number of entries displayed in the main shortcut list. */
-#define MAIN_VISIBLE (((COLS + 40) / 20) * 2)
+#define MAIN_VISIBLE (size_t) (((COLS + 40) / 20) * 2)
 
 /* The minimum editor window columns and rows required for nano to work
  * correctly. */
